Prepend sorts for selected table name when ambiguous with join
Now sorting by any column will work if a JOIN is used with resulting
query. Before, there would be an error from the database about ambiguous
ORDER BY column if sorting by a column that is in more than one of the
JOINed tables. The ORDER BY column is prepended with the selected table
name only when the column's name is ambiguous.