Adds chunk method to query builder to parse sorts
Since 1.17.3 version, when @AlexVanderbist fix duplication of sorts, if you use "chunk" method con query builder, the default sort doesn't affect because the chunk funcion use the enforceOrderBy function before get on queryBuilder and the parseSorts is not executed, so the default model sort is applied (id).

This PR adds the chunk function to ensure parse sorts before enforceOrderBy method is executed on chunk queries.